+++ package.json
@@ -114,6 +114,7 @@
 		"@types/react": "17.0.45",
 		"@types/react-dom": "17.0.16",
 		"@types/wordpress__block-editor": "6.0.6",
+		"@types/wordpress__blocks": "^11.0.1",
 		"@types/wordpress__compose": "4.0.1",
 		"@types/wordpress__data": "4.6.11",
 		"@types/wordpress__data-controls": "2.2.0",
